intial commit for allowing the classic tester to support multiple backends and firrtl interpreter backend implementation
